move props, state, and handlers to EditGameView
